requirements.txt
numpy
scipy
qulacs
tequila-basic
